shell: output: move log output handling to output/log.[ch]
Problem: Log messages from the job shell do not end up in the expected
files when using separate output file per rank or per task.

Move the log handling functionality to output/log.[ch] for better
organization. Tweak the `shell.log` callback so it sends log data
to a local stderr file if one is open, regardless of whether it is
invoked on shell rank 0 or not.

Always initialize `shell.log` plugin callback even if stderr is not
a file now that the code does the right thing in all situations.
